fix:terminal already exists
