+++ frontend/providers/terminal/.env.template
